refactoring combine function to switch argument positions. reuse cliTransform. rename w next page token request transform.
